Create PHP code that creates an HTML table with some sample data
In this code, we first define an array $rows that contains the data we want to display in the table. Each row of the table is represented as an array, and the entire table is stored as an array of rows.

We then use a foreach loop to iterate over each row in the $rows array, and within each row, we use another foreach loop to iterate over each cell in the row. For each cell, we output an HTML <td> element with the cell's contents.

Finally, we wrap the entire table in HTML <table> and <tr> elements to create a valid HTML table.

When this code is executed, it will output an HTML table with the data from the $rows array. You can modify the data in the array to display different information in the table, or modify the HTML markup to change the appearance of the table.